+++ src/main/java/de/swiftbyte/jdaboot/JDABootConfigurationManager.java
@@ -5,10 +5,10 @@ import de.swiftbyte.jdaboot.cli.ConsoleCommandManager;
 import de.swiftbyte.jdaboot.configuration.ConfigProvider;
 import de.swiftbyte.jdaboot.configuration.ConfigProviderChain;
 import de.swiftbyte.jdaboot.configuration.ConfigValueManager;
-import de.swiftbyte.jdaboot.embeds.EmbedManager;
+import de.swiftbyte.jdaboot.embed.EmbedManager;
 import de.swiftbyte.jdaboot.event.EventManager;
-import de.swiftbyte.jdaboot.interactions.button.ButtonManager;
-import de.swiftbyte.jdaboot.interactions.command.CommandManager;
+import de.swiftbyte.jdaboot.interaction.button.ButtonManager;
+import de.swiftbyte.jdaboot.interaction.command.CommandManager;
 import de.swiftbyte.jdaboot.scheduler.SchedulerManager;
 import de.swiftbyte.jdaboot.variables.TranslationProvider;
 import lombok.AccessLevel;
