+++ src/main/java/de/swiftbyte/jdaboot/embed/AdvancedEmbed.java
@@ -1,4 +1,4 @@
-package de.swiftbyte.jdaboot.embeds;
+package de.swiftbyte.jdaboot.embed;
 
 import de.swiftbyte.jdaboot.annotation.embed.Embed;
 import de.swiftbyte.jdaboot.annotation.embed.EmbedField;
