+++ src/main/java/de/swiftbyte/jdaboot/embed/EmbedColor.java
@@ -1,4 +1,4 @@
-package de.swiftbyte.jdaboot.embeds;
+package de.swiftbyte.jdaboot.embed;
 
 import lombok.Getter;
